Enum cached_framework_packages::aptos_stdlib::ScriptFunctionCall

pub enum ScriptFunctionCall {

    AccountCreateAccount {
        auth_key: AccountAddress,
    },
    AccountRotateAuthenticationKey {
        new_auth_key: Vec<u8>,
    },
    AccountUtilsCreateAndFundAccount {
        account: AccountAddress,
        amount: u64,
    },
    ReconfigurationForceReconfigure {},
    StakeAddStake {
        amount: u64,
    },
    StakeIncreaseLockup {
        new_locked_until_secs: u64,
    },
    StakeRegisterValidatorCandidate {
        consensus_pubkey: Vec<u8>,
        network_address: Vec<u8>,
        fullnode_address: Vec<u8>,
    },
    StakeRotateConsensusKey {
        pool_address: AccountAddress,
        new_consensus_pubkey: Vec<u8>,
    },
    StakeSetOperator {
        new_operator: AccountAddress,
    },
    StakeUnlock {
        amount: u64,
    },
    StakeWithdraw {},
    TestCoinClaimMintCapability {},
    TestCoinDelegateMintCapability {
        to: AccountAddress,
    },
    TestCoinMint {
        dst_addr: AccountAddress,
        amount: u64,
    },
    TokenCreateLimitedCollectionScript {
        name: Vec<u8>,
        description: Vec<u8>,
        uri: Vec<u8>,
        maximum: u64,
    },
    TokenCreateLimitedTokenScript {
        collection: Vec<u8>,
        name: Vec<u8>,
        description: Vec<u8>,
        monitor_supply: bool,
        initial_balance: u64,
        maximum: u64,
        uri: Vec<u8>,
        royalty_points_per_million: u64,
    },
    TokenCreateUnlimitedCollectionScript {
        name: Vec<u8>,
        description: Vec<u8>,
        uri: Vec<u8>,
    },
    TokenCreateUnlimitedTokenScript {
        collection: Vec<u8>,
        name: Vec<u8>,
        description: Vec<u8>,
        monitor_supply: bool,
        initial_balance: u64,
        uri: Vec<u8>,
        royalty_points_per_million: u64,
    },
    TokenDirectTransferScript {
        creators_address: AccountAddress,
        collection: Vec<u8>,
        name: Vec<u8>,
        amount: u64,
    },
    TokenInitializeTokenForId {
        creators_address: AccountAddress,
        collection: Vec<u8>,
        name: Vec<u8>,
    },
    TokenInitializeTokenScript {},
    TokenTransfersCancelOfferScript {
        receiver: AccountAddress,
        creator: AccountAddress,
        collection: Vec<u8>,
        name: Vec<u8>,
    },
    TokenTransfersClaimScript {
        sender: AccountAddress,
        creator: AccountAddress,
        collection: Vec<u8>,
        name: Vec<u8>,
    },
    TokenTransfersOfferScript {
        receiver: AccountAddress,
        creator: AccountAddress,
        collection: Vec<u8>,
        name: Vec<u8>,
        amount: u64,
    },
    TransactionPublishingOptionSetModulePublishingAllowed {
        is_allowed: bool,
    },
    VMConfigSetGasConstants {
        global_memory_per_byte_cost: u64,
        global_memory_per_byte_write_cost: u64,
        min_transaction_gas_units: u64,
        large_transaction_cutoff: u64,
        intrinsic_gas_per_byte: u64,
        maximum_number_of_gas_units: u64,
        min_price_per_gas_unit: u64,
        max_price_per_gas_unit: u64,
        max_transaction_size_in_bytes: u64,
        gas_unit_scaling_factor: u64,
        default_account_size: u64,
    },
    ValidatorSetScriptAddValidator {
        _validator_addr: AccountAddress,
    },
    ValidatorSetScriptCreateValidatorAccount {
        _new_account_address: AccountAddress,
        _human_name: Vec<u8>,
    },
    ValidatorSetScriptCreateValidatorOperatorAccount {
        _new_account_address: AccountAddress,
        _human_name: Vec<u8>,
    },
    ValidatorSetScriptRegisterValidatorConfig {
        _validator_address: AccountAddress,
        _consensus_pubkey: Vec<u8>,
        _validator_network_addresses: Vec<u8>,
        _fullnode_network_addresses: Vec<u8>,
    },
    ValidatorSetScriptRemoveValidator {
        _validator_addr: AccountAddress,
    },
    ValidatorSetScriptSetValidatorConfigAndReconfigure {
        _validator_account: AccountAddress,
        _consensus_pubkey: Vec<u8>,
        _validator_network_addresses: Vec<u8>,
        _fullnode_network_addresses: Vec<u8>,
    },
    ValidatorSetScriptSetValidatorOperator {
        _operator_name: Vec<u8>,
        _operator_account: AccountAddress,
    },
    VersionSetVersion {
        major: u64,
    },

}

Structured representation of a call into a known Move script function.

impl ScriptFunctionCall {
    pub fn encode(self) -> TransactionPayload { .. }
    pub fn decode(&TransactionPayload) -> Option<ScriptFunctionCall> { .. }
}

Variants

AccountCreateAccount

Fields

auth_key: AccountAddress

Basic account creation methods.

AccountRotateAuthenticationKey

Fields

new_auth_key: Vec<u8>

AccountUtilsCreateAndFundAccount

Fields

account: AccountAddress

amount: u64

ReconfigurationForceReconfigure

Fields

Force an epoch change.

StakeAddStake

Fields

amount: u64

Add amount of coins from the account owning the StakePool.

StakeIncreaseLockup

Fields

new_locked_until_secs: u64

Similar to increase_lockup_with_cap but will use ownership capability from the signing account.

StakeRegisterValidatorCandidate

Fields

consensus_pubkey: Vec<u8>

network_address: Vec<u8>

fullnode_address: Vec<u8>

Initialize the validator account and give ownership to the signing account.

StakeRotateConsensusKey

Fields

pool_address: AccountAddress

new_consensus_pubkey: Vec<u8>

Rotate the consensus key of the validator, it’ll take effect in next epoch.

StakeSetOperator

Fields

new_operator: AccountAddress

Allows an owner to change the operator of the stake pool.

StakeUnlock

Fields

amount: u64

Similar to unlock_with_cap but will use ownership capability from the signing account.

StakeWithdraw

Fields

Withdraw from account’s inactive stake.

TestCoinClaimMintCapability

Fields

Claim the delegated mint capability and destroy the delegated token.

TestCoinDelegateMintCapability

Fields

to: AccountAddress

Create delegated token for the address so the account could claim MintCapability later.

TestCoinMint

Fields

dst_addr: AccountAddress

amount: u64

Create new test coins and deposit them into dst_addr’s account.

TokenCreateLimitedCollectionScript

Fields

name: Vec<u8>

description: Vec<u8>

uri: Vec<u8>

maximum: u64

TokenCreateLimitedTokenScript

Fields

collection: Vec<u8>

name: Vec<u8>

description: Vec<u8>

monitor_supply: bool

initial_balance: u64

maximum: u64

uri: Vec<u8>

royalty_points_per_million: u64

TokenCreateUnlimitedCollectionScript

Fields

name: Vec<u8>

description: Vec<u8>

uri: Vec<u8>

TokenCreateUnlimitedTokenScript

Fields

collection: Vec<u8>

name: Vec<u8>

description: Vec<u8>

monitor_supply: bool

initial_balance: u64

uri: Vec<u8>

royalty_points_per_million: u64

TokenDirectTransferScript

Fields

creators_address: AccountAddress

collection: Vec<u8>

name: Vec<u8>

amount: u64

TokenInitializeTokenForId

Fields

creators_address: AccountAddress

collection: Vec<u8>

name: Vec<u8>

TokenInitializeTokenScript

Fields

TokenTransfersCancelOfferScript

Fields

receiver: AccountAddress

creator: AccountAddress

collection: Vec<u8>

name: Vec<u8>

TokenTransfersClaimScript

Fields

sender: AccountAddress

creator: AccountAddress

collection: Vec<u8>

name: Vec<u8>

TokenTransfersOfferScript

Fields

receiver: AccountAddress

creator: AccountAddress

collection: Vec<u8>

name: Vec<u8>

amount: u64

TransactionPublishingOptionSetModulePublishingAllowed

Fields

is_allowed: bool

VMConfigSetGasConstants

Fields

global_memory_per_byte_cost: u64

global_memory_per_byte_write_cost: u64

min_transaction_gas_units: u64

large_transaction_cutoff: u64

intrinsic_gas_per_byte: u64

maximum_number_of_gas_units: u64

min_price_per_gas_unit: u64

max_price_per_gas_unit: u64

max_transaction_size_in_bytes: u64

gas_unit_scaling_factor: u64

default_account_size: u64

ValidatorSetScriptAddValidator

Fields

_validator_addr: AccountAddress

ValidatorSetScriptCreateValidatorAccount

Fields

_new_account_address: AccountAddress

_human_name: Vec<u8>

ValidatorSetScriptCreateValidatorOperatorAccount

Fields

_new_account_address: AccountAddress

_human_name: Vec<u8>

ValidatorSetScriptRegisterValidatorConfig

Fields

_validator_address: AccountAddress

_consensus_pubkey: Vec<u8>

_validator_network_addresses: Vec<u8>

_fullnode_network_addresses: Vec<u8>

ValidatorSetScriptRemoveValidator

Fields

_validator_addr: AccountAddress

ValidatorSetScriptSetValidatorConfigAndReconfigure

Fields

_validator_account: AccountAddress

_consensus_pubkey: Vec<u8>

_validator_network_addresses: Vec<u8>

_fullnode_network_addresses: Vec<u8>

ValidatorSetScriptSetValidatorOperator

Fields

_operator_name: Vec<u8>

_operator_account: AccountAddress

VersionSetVersion

Fields

major: u64

Updates the major version to a larger version.

Implementations

impl ScriptFunctionCall

pub fn encode(self) -> TransactionPayload

Build an Aptos TransactionPayload from a structured object ScriptFunctionCall.

pub fn decode(payload: &TransactionPayload) -> Option<ScriptFunctionCall>

Try to recognize an Aptos TransactionPayload and convert it into a structured object ScriptFunctionCall.